requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Mail==0.9.1
email-validator==2.0.0
python-dotenv==1.0.0
bcrypt==4.0.1
itsdangerous==2.1.2
Pillow==9.5.0
APScheduler==3.10.1
PyJWT==2.6.0
pyotp==2.8.0
qrcode==7.4.2 